Server: Remove ATTR_PERSISTENT=true
I ran into a connection limit because our Apache in beta has more
than 10 mod_php workers, but the mysql user had a limit of 10
concurrent connections. I've fixed that by raising the mysql
user connection limit in Beta, but generally it seems not worth
the complexity to re-use connections on a service that has on
average <1 requests per minute.

Change-Id: Ic78c7123db2642b9c6987b9bc1c22310267d1f83